+++ packages/dashpub_server/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+- Added a documentation.
+
 ## 0.0.1
 
 - Initial release of the Dashpub Server.